Refactor monetized attributes tests into a dedicated file
